+++ assets/src/js/containers/App/app.js
@@ -1,18 +1,19 @@
+/* eslint-disable object-curly-newline */
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
-import Title from '../../components/Title';
 import Home from '../../components/Home';
 import Calculator from '../../components/Calculator';
 import Pomodoro from '../../components/Pomodoro';
+/* eslint-enable object-curly-newline */
 
 class App extends Component {
   render() {
-    const text = 'Project Orbital';
+    const text = 'Welcome to Project Orbital';
     return (
       <Router>
         <div>
           <nav className="navbar">
-            <div className="logo">Welcome to Project Orbital</div>
+            <div className="logo">{text}</div>
             <ul className="navbar-nav">
               <li><Link to={'/'} className="nav-link"> Home </Link></li>
               <li><Link to={'/calculator'} className="nav-link"> Calculator </Link></li>